// apps/api/src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export namespace Models {
  export type public_Product = {
    description: CodecTypes['pg/text@1']['output'] | null;
    id: CodecTypes['pg/int4@1']['output'];
    name: CodecTypes['pg/text@1']['output'];
    price: CodecTypes['pg/numeric@1']['output'];
    stock: CodecTypes['pg/int4@1']['output'];
    readonly [RelationKeys]?: never;
  };
}